+++ src/InvalidRef.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Swaggest\JsonSchema;
+
+class InvalidRef extends InvalidValue
+{
+
+}